fix: Typo in Profile folder, Rename it
