+++ src/Components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Navigation from './Navigation/Navigation';
-import Profile from './profile/Profile';
+import Profile from './Profile/Profile';
 import './App.css';
 
 class App extends React.Component {
