+++ src/Components/Profile/Summary/Icon/Icon.js
