+++ src/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './Components/App';
 
 ReactDOM.render(<App />, document.querySelector("#root"));
